% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countIssues.R
\name{activeIssues}
\alias{activeIssues}
\title{Create a data frame with the distribution of active issues in the 5 components, by month
for the last 6 months}
\usage{
activeIssues(df, ds = Sys.Date(), p = 6)
}
\arguments{
\item{df}{data frame}

\item{ds}{date start; string - format YYYY-MM-DD}

\item{p}{period; int}
}
\value{
data frame of active issues by component in the last 6 months
}
\description{
Create a data frame with the distribution of active issues in the 5 components, by month
for the last 6 months
}
