% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{aiPlot}
\alias{aiPlot}
\title{Plot active epics by components by month for the last 6 months}
\usage{
aiPlot(df)
}
\arguments{
\item{df}{data frame}
}
\value{
plotly bar graph
}
\description{
Plot active epics by components by month for the last 6 months
}
