% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{cePlot}
\alias{cePlot}
\title{Plot the number of closed epics per month for the last 12 months}
\usage{
cePlot(df)
}
\arguments{
\item{df}{data frame}
}
\value{
plotly bar graph
}
\description{
Plot the number of closed epics per month for the last 12 months
}
