% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countIssues.R
\name{closedEpics}
\alias{closedEpics}
\title{Count the number of closed epics per month for the last 12 months}
\usage{
closedEpics(df, rd = Sys.Date(), p = 12)
}
\arguments{
\item{df}{data frame}

\item{rd}{resolved date; string - format YYYY-MM-DD}

\item{p}{period; int}
}
\value{
data frame with number of closed epics in the last year
}
\description{
Create a data frame with the number of closed epics per month
}
