% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countIssues.R
\name{countIssues}
\alias{countIssues}
\title{Count the number of issues}
\usage{
countIssues(
  df,
  filt = list("Status"),
  filt_val = list(list("To Do", "In Progress", "Done")),
  group,
  piv = list(FALSE, "")
)
}
\arguments{
\item{df}{data frame}

\item{filt}{a list of column names to filter df with}

\item{filt_val}{a list of corresponding values being filtered for each column in f}

\item{group}{a list of column names}

\item{piv}{optional pivot of the data frame (a list pair: (Boolean, column name (string)))}
}
\value{
data frame with counts for the query
}
\description{
Allows the user to filter and group the JIRA metrics however they like, and then returns the
number of issues for their query
}
