% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countIssues.R
\name{divIssues}
\alias{divIssues}
\title{Create a data frame with the distribution of issues by division, grouped by field}
\usage{
divIssues(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame of issues by division, grouped by field
}
\description{
Create a data frame with the distribution of issues by division, grouped by field
}
