% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{divPlot}
\alias{divPlot}
\title{Plot distribution of epics by division, grouped by field}
\usage{
divPlot(df)
}
\arguments{
\item{df}{data frame}
}
\value{
plotly bar graph
}
\description{
Plot distribution of epics by division, grouped by field
}
