% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira-metrics.R
\docType{package}
\name{jiraR-package}
\alias{jiraR-package}
\alias{jira-metrics}
\title{JIRA R Package}
\description{
Create a package that summarizes and plots the following metrics:
 - Number of closed epics per month for the last 12 months.
 - Number of new issues each month for the last 12 months.
 - Distribution of active issues in the 5 components, by month for the last 6 months.
 - Distribution of issues by division, grouped by field (likely a stacked bar chart).
 - Distribution of issues by workspaces.
}
