% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countIssues.R
\name{missingData}
\alias{missingData}
\title{Subset the data to find all missing values}
\usage{
missingData(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame that contains all the rows with NA
}
\description{
Subset the data to find all missing values
}
