% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countIssues.R
\name{newIssues}
\alias{newIssues}
\title{Count the number of new issues per month for the last 12 months}
\usage{
newIssues(df, ds = Sys.Date(), p = 12)
}
\arguments{
\item{df}{data frame}

\item{ds}{date start; string - format YYYY-MM-DD}

\item{p}{period; int}
}
\value{
data frame with number of new issues in the last year
}
\description{
Create a data frame with the number of closed epics per month.
Define new issues as Status = To Do (Can change this to be Status = To Do and In Progress)
}
