% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{niPlot}
\alias{niPlot}
\title{Plot the number of new epics per month for the last 12 months}
\usage{
niPlot(df)
}
\arguments{
\item{df}{data frame}
}
\value{
plotly bar graph
}
\description{
Plot the number of new epics per month for the last 12 months
}
