% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countIssues.R
\name{sumTable}
\alias{sumTable}
\title{A summary table of the issue counts in the last 12 months}
\usage{
sumTable(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame with summary of closed epics and new issues in the last year
}
\description{
Create a data frame with the number of closed epics and new issues per month.
}
