% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countIssues.R
\name{wsIssues}
\alias{wsIssues}
\title{Create a data frame with the distribution of issues by workspaces}
\usage{
wsIssues(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame of the number of issues by workspace
}
\description{
Create a data frame with the distribution of issues by workspaces
}
