% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetrics.R
\name{wsPlot}
\alias{wsPlot}
\title{Plot distribution of epics by workspaces}
\usage{
wsPlot(df)
}
\arguments{
\item{df}{data frame}
}
\value{
plotly bar graph
}
\description{
Plot distribution of epics by workspaces
}
