% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmltodf.R
\name{xmltodf}
\alias{xmltodf}
\title{Convert an XML file to a Data Frame}
\usage{
xmltodf(xml_file)
}
\arguments{
\item{xml_file}{data frame}
}
\value{
data frame
}
\description{
Convert an XML file to a Data Frame
}
